BE/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=1.10.0
cloudinary>=1.27.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.9